requirements.txt
Django
djangorestframework
drf-spectacular
django-cors-headers
psycopg2-binary
dj-database-url
python-dotenv
django-extensions
Werkzeug
Faker
Cython
gunicorn
django-grappelli
whitenoise